requirements.txt
flask==2.3.3
pandas==2.2.3
requests==2.31.0
werkzeug==2.3.7
python-dotenv==1.0.0
simple-salesforce==1.12.4
xlrd
openpyxl==3.1.3
